chore(styles): renamed file to be dash-case
